introduce attribute java package and move all related code there.
