+++ henson-plugin/src/main/resources/build.properties
@@ -1,2 +1,2 @@
-#Tue Dec 12 09:15:47 PST 2017
+#Tue Dec 12 09:20:41 PST 2017
 dart.version=3.0.0-RC2-SNAPSHOT
